+++ src/main/kotlin/blackjack/model/card/Card.kt
@@ -1,8 +1,5 @@
 package blackjack.model.card
 
-import blackjack.model.Rank
-import blackjack.model.Suit
-
 data class Card(val suit: Suit, val rank: Rank) {
     fun isAce() = (rank == Rank.ACE)
 }
