+++ src/test/kotlin/blackjack/model/card/CardTest.kt
@@ -1,7 +1,6 @@
 package blackjack.model.card
 
 import blackjack.fixture.createCard
-import blackjack.model.Rank
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
