+++ src/test/kotlin/blackjack/model/participant/PlayerTest.kt
@@ -2,7 +2,7 @@ package blackjack.model.participant
 
 import blackjack.fixture.createCard
 import blackjack.fixture.createPlayer
-import blackjack.model.Rank
+import blackjack.model.card.Rank
 import blackjack.model.card.Hand
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
